// src/types/errorMessages.ts
export enum ErrorMessages {
  none = '',
  titleIsEmpty = 'Title should not be empty',
  todosLoadError = 'Unable to load todos',
  todoAddError = 'Unable to add a todo',
  todoDeleteError = 'Unable to delete a todo',
  todosDeleteError = 'Unable to delete todos',
  todosUpdateError = `Unable to update a todo`,
}